--- src/main/java/kitchenpos/order/application/TableService.java
@@ -1,59 +1,0 @@
-package kitchenpos.order.application;
-
-import kitchenpos.exception.EntityNotFoundException;
-import kitchenpos.order.domain.*;
-import kitchenpos.order.dto.OrderTableRequest;
-import kitchenpos.order.dto.OrderTableResponse;
-import kitchenpos.order.exception.OrderTableExceptionCode;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class TableService {
-    private final OrderRepository orderRepository;
-    private final OrderTableRepository orderTableRepository;
-
-    public TableService(OrderRepository orderRepository, OrderTableRepository orderTableRepository) {
-        this.orderRepository = orderRepository;
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    @Transactional
-    public OrderTableResponse create(OrderTableRequest request) {
-        return OrderTableResponse.of(orderTableRepository.save(request.toOrderTable()));
-    }
-
-    @Transactional(readOnly = true)
-    public List<OrderTableResponse> list() {
-        return OrderTableResponse.list(orderTableRepository.findAll());
-    }
-
-    @Transactional
-    public OrderTableResponse changeEmpty(Long orderTableId, OrderEmpty request) {
-        OrderTable savedOrderTable = findOrderTableById(orderTableId);
-        List<Order> orders = findAllByOrderTableId(savedOrderTable.getId());
-
-        savedOrderTable.changeEmpty(request.isEmpty(), orders);
-
-        return OrderTableResponse.of(orderTableRepository.save(savedOrderTable));
-    }
-
-    private OrderTable findOrderTableById(Long orderTableId) {
-        return orderTableRepository.findById(orderTableId)
-                .orElseThrow(() -> new EntityNotFoundException(OrderTableExceptionCode.NOT_FOUND_BY_ID.getMessage()));
-    }
-
-    private List<Order> findAllByOrderTableId(Long orderTableId) {
-        return orderRepository.findAllByOrderTableId(orderTableId);
-    }
-
-    @Transactional
-    public OrderTableResponse changeNumberOfGuests(Long orderTableId, OrderGuests request) {
-        OrderTable orderTable = findOrderTableById(orderTableId);
-        orderTable.changeNumberOfGuests(request.getNumberOfGuests());
-
-        return OrderTableResponse.of(orderTableRepository.save(orderTable));
-    }
-}
