+++ src/main/java/kitchenpos/tablegroup/ui/TableGroupRestController.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.ui;
+package kitchenpos.tablegroup.ui;
 
-import kitchenpos.order.application.TableGroupService;
-import kitchenpos.order.dto.TableGroupRequest;
-import kitchenpos.order.dto.TableGroupResponse;
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
